Refactorizar el esquema de la base de datos para utilizar INT en lugar de BIGINT para las claves principales en todas las tablas. Actualizar la configuración de conexión en DAO para utilizar el usuario "root" y un nuevo nombre de base de datos. Cambiar el controlador de la base de datos de mysql.connector a pymysql en ProductoDAO. Realizar pequeños ajustes en requirements.txt para eliminar dependencias innecesarias.

+++ database.sql
@@ -1,5 +1,5 @@
 CREATE TABLE productos (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   nombre VARCHAR(255) NOT NULL,
   descripcion TEXT,
   precio DECIMAL(10, 2) NOT NULL,
@@ -7,8 +7,8 @@ CREATE TABLE productos (
 );
 
 CREATE TABLE entradas_inventario (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  producto_id BIGINT,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  producto_id INT,
   cantidad INT NOT NULL,
   fecha TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (producto_id) REFERENCES productos(id)
@@ -23,59 +23,59 @@ CREATE TABLE salidas_inventario (
 );
 
 CREATE TABLE usuarios (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   nombre_usuario VARCHAR(255) UNIQUE NOT NULL,
   contrasena VARCHAR(255) NOT NULL,
   rol VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE roles_permisos (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   rol VARCHAR(50) NOT NULL,
   permiso VARCHAR(50) NOT NULL
 );
 
 CREATE TABLE alertas_inventario (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  producto_id BIGINT,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  producto_id INT,
   nivel_alerta INT NOT NULL,
   fecha TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (producto_id) REFERENCES productos(id)
 );
 
 CREATE TABLE proveedores (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   nombre VARCHAR(255) NOT NULL,
   contacto VARCHAR(255),
   direccion TEXT
 );
 
 CREATE TABLE categorias_productos (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   nombre VARCHAR(255) NOT NULL,
   descripcion TEXT
 );
 
 CREATE TABLE historial_precios (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  producto_id BIGINT,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  producto_id INT,
   precio DECIMAL(10, 2) NOT NULL,
   fecha TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (producto_id) REFERENCES productos(id)
 );
 
 CREATE TABLE pedidos (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  proveedor_id BIGINT,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  proveedor_id INT,
   fecha_pedido TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   estado VARCHAR(50) NOT NULL,
   FOREIGN KEY (proveedor_id) REFERENCES proveedores(id)
 );
 
 CREATE TABLE detalles_pedidos (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  pedido_id BIGINT,
-  producto_id BIGINT,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  pedido_id INT,
+  producto_id INT,
   cantidad INT NOT NULL,
   precio_unitario DECIMAL(10, 2) NOT NULL,
   FOREIGN KEY (pedido_id) REFERENCES pedidos(id),
